MDRun now returns both complex and a bool.
When MDRun is created with image_after set to false, all_atoms_in_box
is always True. When MDRun is created with image_after set to True,
then after imaging, FixBox will report if the protein is outside the
edges of the box. Then, an error will be logged, but the protocol will
continue.