Fix bug in run_tleap()
When assigning chainIDs to the parmed object read from the
pre_{name}.prmtop and pre_{name}.rst7 generated by tleap, may have more
or less residues than the original {name}.pdb that tleap read in. To
account for this, we iterate only over the residues from the parmed
object and in case there're not enough residues from the original PDB
from wich to get chainIDs, we just reuse the chainID read from the
last residue.