Fix error preventing restart from NVT
energy file was getting its default name (ener.edr) and when the
protocol was terminated during NVT, the protocol would restart
from the minimization, which would find a file called 'ener.edr'
but that was coming from the NVT and so it would complain about a
checksum error. Now the energy file gets a proper name.